Add License
In general:

Use this open source project as an example to build your own.
We don't really care if you attribute us for the software setup itself.
However it is always apprecaited ^_^
What we do care about is if you use our content.
All content is copyright 2017 Sacred Seed Studio